requirements.txt
fastapi-users-db-sqlalchemy
fastapi
fastapi-users[sqlalchemy2]
uvicorn[standard]
aiosqlite